Corrects dice manager and pearson crytery

+++ Лабораторная работа 1/DiceManager.h
@@ -12,8 +12,8 @@ public:
 	DiceManager(Dice*, unsigned int);
 	void setDisplayFunc(void (*callback)(int)); //  функция обратного вызова
 	void testDice();
-	unsigned int getDropped(size_t face);
-	unsigned int getExpectedDrop(size_t face) {return dice->getProbability(face) * testCount;} 
+	float getDropped(size_t face);
+	float getExpectedDrop(size_t face) {return dice->getProbability(face) * testCount;} 
 	float getEProbability(size_t); // эмпирическая вероятность
 	double pearson();
 	~DiceManager(void);
